third_party/blink/web_tests/svg/custom/keySplines-parsing-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests parsing of keySplines attribute.");

var animate = document.createElementNS("http://www.w3.org/2000/svg", "animate");
animate.setAttribute("keySplines", ";;");
animate.setAttribute("keySplines", "0 ,0  1 , 1  ;;");

var successfullyParsed = true;
</script>
</body>
</html>
