third_party/blink/web_tests/svg/custom/keySplines-parsing-error-expected.txt

CONSOLE ERROR: line 11: Error: <animate> attribute keySplines: Invalid value, ";;".
CONSOLE ERROR: line 12: Error: <animate> attribute keySplines: Invalid value, "0 ,0  1 , 1  ;;".
Tests parsing of keySplines attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
